Refactoring Home page to be named Home page and not profile page
