+++ sudokuru/app/Pages/ProfilePage.tsx
@@ -5,7 +5,7 @@ import {StyleSheet, View} from "react-native";
 import {Text} from 'react-native-paper';
 import SaveGamePreferencesToggle from "../Components/Profile/SaveGamePreferencesToggle";
 import StatisticsButton from "../Components/Statistics/StatisticsButton";
-import HomeButton from "../Components/Landing/HomeButton";
+import HomeButton from "../Components/Home/HomeButton";
 import SidebarMenu from "../Components/SidebarMenu";
 
 const ProfilePage = () => {
